Refectoring Application for apply 'Redux Toolkit'
Applied 'Redux Toolkit' through the codesoom lecture!
