+++ package.json
@@ -44,12 +44,12 @@
     "@codeceptjs/configure": "^0.5.2",
     "@emotion/core": "^10.1.1",
     "@emotion/styled": "^10.0.27",
+    "@reduxjs/toolkit": "^1.4.0",
     "emotion": "^10.0.27",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-redux": "^7.2.0",
     "react-router-dom": "^5.2.0",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0"
+    "redux": "^4.0.5"
   }
 }
